MDL-64152 library: Use servername from configuration for LTI.
Both sides of OAuth for LTI need to agree on a servername. If an app server
is run behind a proxy the hostname is not equal to the external servername.
Variables HTTP_ORIGIN, HTTP_X_FORWARDED_SERVER_PORT, and HTTP_HOST might
not be present and are not trustworthy as they are user provided.
This change uses the configured servername `$CFG->wwwroot`.